Create Restart-FailedSpooler
This cmdlet was created to run automatically through Windows Task Scheduler to eliminate the need for a user to submit a ticket whenever the print spooler needs to be restarted. In the Applications and Service Event Log PrintService - Admin (Microsoft-Windows-PrintService/Admin) Event ID 372, 350, and 314 should trigger this script to run. I have not been able to get the Invoke-Command to execute through Task scheduler. It works if the script is run normally but not the automatic way which I believe is a security feature. If you solve this issue please let me know.